requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2.2
django-allauth==0.51.0
django-bootstrap-datepicker-plus==5.0.3
django-bootstrap4==22.3
django-widget-tweaks==1.4.12
idna==3.4
oauthlib==3.2.2
Pillow==9.5.0
pycparser==2.21
pydantic==1.10.9
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==2.0.3
